working, changed null ID to 0 but breaks if context handle is returned instead of root handle in entityContext_open

+++ math/internal/units/constants.c
@@ -1,4 +1,4 @@
 #include "units/constants.h"
 #include <math.h>
 
-const float OCT_2PI_f = 2 * M_PI;
+const float OCT_2PI_f = 2 * (float)M_PI;
